Move the Result struct to result.go and output functions to output.go, so that it is easier to copy that code between the examples
